SZLABAN/requirements.txt
wheel
Flask
requests
rich
smbus2
spidev
rpi-gpio
rpi-lgpio